@@ -26,18 +26,18 @@
         "source": "https://github.com/TheDragonCode/support"
     },
     "require": {
-        "php": "^7.4 || ^8.0",
+        "php": "^8.0",
+        "ext-bcmath": "*",
         "ext-ctype": "*",
         "ext-dom": "*",
         "ext-json": "*",
         "ext-mbstring": "*",
         "dragon-code/contracts": "^2.15",
         "psr/http-message": "^1.0.1",
-        "symfony/deprecation-contracts": "^2.5 || ^3.0",
-        "voku/portable-ascii": "^1.4.8|^2.0"
+        "voku/portable-ascii": "^1.4.8 || ^2.0"
     },
     "require-dev": {
-        "ext-bcmath": "*",
+        "illuminate/contracts": "^9.7",
         "phpunit/phpunit": "^8.0 || ^9.0",
         "symfony/var-dumper": "^5.2 || ^6.0"
     },
@@ -45,7 +45,6 @@
         "andrey-helldar/support": "*"
     },
     "suggest": {
-        "ext-bcmath": "Require the extension if you will be using DragonCode\\Support\\Facades\\Helpers\\Digit.",
         "dragon-code/laravel-support": "Various helper files for the Laravel and Lumen frameworks",
         "symfony/thanks": "Give thanks (in the form of a GitHub) to your fellow PHP package maintainers"
     },
