@@ -25,7 +25,7 @@ use DragonCode\Support\Tools\Replace as Tool;
  */
 class Replace extends Facade
 {
-    protected static function getFacadeAccessor()
+    protected static function getFacadeAccessor(): string
     {
         return Tool::class;
     }
